Add export functions to index file
